api/requirements.txt
flask==2.2.3
gunicorn==20.1.0
paho-mqtt==1.6.1
python-dotenv==1.0.0
flask-cors==3.0.10